% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_rename.R
\name{d_rename}
\alias{d_rename}
\title{d_rename}
\usage{
d_rename(df, x, row = FALSE, col = TRUE)
}
\arguments{
\item{x}{x is vector or list of names you wish put in the data frame. row and col args is there for selection.}
}
\value{
returns data frame with renamed columns names or row names.
}
\description{
Renames a columns or rows of a data frame
}
