% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dclean.R
\name{dclean}
\alias{dclean}
\title{dclean}
\usage{
dclean(df, replace = NA, exchange = input)
}
\arguments{
\item{x}{replace chooses NA values while exchange changes the values. input allowed is AVG, MED, etc}
}
\value{
AVG for average the columns, MED for median of columns. optional input like a number or a string
}
\description{
Cleans the NA values in data frame and changes its value
}
