% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dclean_text.R
\name{dclean_text}
\alias{dclean_text}
\title{dclean_text}
\usage{
dclean_text(vector, input)
}
\arguments{
\item{x}{input allow to put in a string you want clean}
}
\value{
vector fully cleaned
}
\description{
Clean up text.
}
