% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dformat.R
\name{dformat}
\alias{dformat}
\title{dformat}
\usage{
dformat(df, format_type = input)
}
\arguments{
\item{x}{format_type arg choose between date and number for a change of format}
}
\value{
new date or number format
}
\description{
Change the format of numbers and date to something r can interpret.
}
